+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-[Unreleased]
+[1.10.0] - 2020-05-22
 ---------------------
 ##### Changed
 - In OC mode appearance errors are no output as warnings.